+++ command/operator_init.go
@@ -1,13 +1,11 @@
 package command
 
 import (
-	"encoding/json"
 	"fmt"
 	"net/url"
 	"runtime"
 	"strings"
 
-	"github.com/ghodss/yaml"
 	"github.com/hashicorp/vault/api"
 	"github.com/hashicorp/vault/helper/pgpkeys"
 	"github.com/mitchellh/cli"
@@ -53,7 +51,7 @@ func (c *OperatorInitCommand) Help() string {
 Usage: vault operator init [options]
 
   Initializes a Vault server. Initialization is the process by which Vault's
-  storage backend is prepared to receive data. Since Vault server's share the
+  storage backend is prepared to receive data. Since Vault servers share the
   same storage backend in HA mode, you only need to initialize one Vault to
   initialize the storage backend.
 
@@ -405,7 +403,7 @@ func (c *OperatorInitCommand) consulAuto(client *api.Client, req *api.InitReques
 		// requiring the client to update VAULT_ADDR and to run init again.
 		c.UI.Output(wrapAtLength(fmt.Sprintf(
 			"Discovered %d uninitialized Vault servers with Consul service name "+
-				"%q. To initialize these Vatuls, set any one of the following "+
+				"%q. To initialize these Vaults, set any one of the following "+
 				"environment variables and run \"vault init\":",
 			len(uninitedVaults), c.flagConsulService)))
 		c.UI.Output("")
@@ -434,15 +432,10 @@ func (c *OperatorInitCommand) init(client *api.Client, req *api.InitRequest) int
 		return 2
 	}
 
-	switch c.flagFormat {
-	case "yaml", "yml":
-		return c.initOutputYAML(req, resp)
-	case "json":
-		return c.initOutputJSON(req, resp)
+	switch Format(c.UI) {
 	case "table":
 	default:
-		c.UI.Error(fmt.Sprintf("Unknown format: %s", c.flagFormat))
-		return 1
+		return OutputData(c.UI, newMachineInit(req, resp))
 	}
 
 	for i, key := range resp.Keys {
@@ -466,8 +459,8 @@ func (c *OperatorInitCommand) init(client *api.Client, req *api.InitRequest) int
 	if req.StoredShares < 1 {
 		c.UI.Output("")
 		c.UI.Output(wrapAtLength(fmt.Sprintf(
-			"Vault initialized with %d key shares an a key threshold of %d. Please "+
-				"securely distributed the key shares printed above. When the Vault is "+
+			"Vault initialized with %d key shares and a key threshold of %d. Please "+
+				"securely distribute the key shares printed above. When the Vault is "+
 				"re-sealed, restarted, or stopped, you must supply at least %d of "+
 				"these keys to unseal it before it can start servicing requests.",
 			req.SecretShares,
@@ -501,26 +494,6 @@ func (c *OperatorInitCommand) init(client *api.Client, req *api.InitRequest) int
 	}
 
 	return 0
-}
-
-// initOutputYAML outputs the init output as YAML.
-func (c *OperatorInitCommand) initOutputYAML(req *api.InitRequest, resp *api.InitResponse) int {
-	b, err := yaml.Marshal(newMachineInit(req, resp))
-	if err != nil {
-		c.UI.Error(fmt.Sprintf("Error marshaling YAML: %s", err))
-		return 2
-	}
-	return PrintRaw(c.UI, strings.TrimSpace(string(b)))
-}
-
-// initOutputJSON outputs the init output as JSON.
-func (c *OperatorInitCommand) initOutputJSON(req *api.InitRequest, resp *api.InitResponse) int {
-	b, err := json.Marshal(newMachineInit(req, resp))
-	if err != nil {
-		c.UI.Error(fmt.Sprintf("Error marshaling JSON: %s", err))
-		return 2
-	}
-	return PrintRaw(c.UI, strings.TrimSpace(string(b)))
 }
 
 // status inspects the init status of vault and returns an appropriate error